changed folder structure; some cleanup
