--- four-point_prediction.csv
@@ -1,77 +1,0 @@
-Time,intensity1,intensity2,intensity3,intensity4,force1,force2,force3,force4
-0,7621,67879,50303,6881,1.15132912,1.353796,1.29474064,1.14884272
-0.1,7439,67700,50159,6869,1.14924932,1.35172628,1.29278852,1.14733412
-0.2,7117,67155,49515,6794,1.14366316,1.34539084,1.28612044,1.14257788
-0.3,7400,66555,49705,6955,1.1447106,1.3434714,1.2868554,1.1432154
-0.4,7219,66816,49829,6544,1.14351456,1.34376048,1.28668416,1.14124656
-0.5,7306,67673,50157,6931,1.14851844,1.35135156,1.2924978,1.14725844
-0.6,7307,67556,50123,6520,1.14692856,1.3493652,1.29079032,1.14428424
-0.7,7522,67836,50564,6963,1.15156732,1.35422236,1.29618844,1.14968908
-0.8,7755,70623,52482,7833,1.16884492,1.3800814,1.31912764,1.169107
-0.9,8113,72253,54025,8385,1.18164352,1.39715392,1.33590784,1.18255744
-1,9746,84352,61274,13470,1.26115784,1.511834,1.43429192,1.27367048
-1.1,9587,87262,63129,13884,1.2748804,1.5358684,1.45478152,1.28931832
-1.2,11221,94997,68352,16304,1.32868472,1.61017208,1.52064488,1.3457636
-1.3,11349,96218,69584,17512,1.33987556,1.6250354,1.53554516,1.36058324
-1.4,11297,96831,69772,17124,1.34072608,1.62812032,1.53720208,1.3603048
-1.6,11523,96885,70397,17072,1.34390796,1.63072428,1.5417246,1.3625526
-1.7,11595,96931,70160,17784,1.345834,1.63256296,1.5426124,1.36662904
-1.8,11568,96639,70441,17422,1.34460728,1.63044584,1.54242056,1.36427672
-1.9,11757,96485,70591,17422,1.34576772,1.6304538,1.54344996,1.36480212
-2,11590,96582,70313,17465,1.3443404,1.62991352,1.54164968,1.3640804
-2.1,11799,98580,72178,18397,1.359254,1.65083816,1.56212744,1.38142328
-2.2,13078,102912,75144,19759,1.3917782,1.69362044,1.60031996,1.41422636
-2.3,12994,104756,75647,21195,1.40200112,1.71032144,1.6125152,1.42955648
-2.4,13474,107471,77604,22606,1.42225284,1.73808276,1.63772964,1.45293636
-2.5,13828,108048,77894,21531,1.42385692,1.74043612,1.63911868,1.449739
-2.6,14085,107967,78231,22206,1.42809436,1.74353788,1.64362492,1.45538092
-2.7,13801,107858,78145,22517,1.426663,1.74269452,1.64285884,1.45594876
-2.8,13768,108179,78774,21827,1.4271968,1.74441776,1.64561696,1.45427504
-3,13855,107996,78874,22073,1.42819912,1.74451288,1.64666296,1.4558116
-3.1,13654,106681,77713,21435,1.41810916,1.73067988,1.6333474,1.44425332
-3.2,14109,107174,78377,21680,1.42491184,1.73761024,1.64085232,1.4503504
-3.3,14032,106353,78135,21924,1.42210848,1.73230704,1.63749456,1.4486256
-3.4,13969,109043,80411,23535,1.44039656,1.7598452,1.66364168,1.47253832
-3.5,14774,115468,84223,26250,1.48217124,1.82050308,1.71551988,1.5207306
-3.7,14772,115577,84167,27011,1.4844706,1.8231754,1.7176378,1.52559364
-3.8,15030,115996,84464,27358,1.48908912,1.82833488,1.72238736,1.5305112
-3.9,14750,115755,84609,27003,1.48607228,1.82544908,1.72079852,1.52724236
-4,15083,116485,85988,26994,1.49410088,1.8348116,1.73234168,1.53412184
-4.1,15262,116565,86252,26820,1.49569348,1.83607156,1.73421988,1.53452836
-4.2,15308,116865,85627,26715,1.49475748,1.835989,1.73102932,1.533085
-4.4,15513,117327,87242,27259,1.50347212,1.84556716,1.74448156,1.54293868
-4.5,15682,116900,85711,27160,1.49867804,1.83877052,1.73397548,1.53724412
-4.6,15488,116940,85792,27286,1.49817672,1.83905544,1.73439816,1.537818
-4.7,14962,114601,84796,26451,1.48307272,1.81785976,1.71771496,1.52167576
-4.8,14757,114537,84707,26534,1.48160292,1.81686372,1.71663492,1.52117364
-5,15040,114684,85223,25807,1.48317576,1.8179796,1.71899064,1.51935288
-5.1,14814,114862,85076,26083,1.48264644,1.81880772,1.71872676,1.52051028
-5.2,15018,114306,84369,26741,1.48219304,1.81580072,1.7152124,1.52158232
-5.4,15020,114156,85057,26315,1.48252352,1.81562048,1.71784784,1.52047472
-5.5,14864,115009,85346,25999,1.48390216,1.82038936,1.72072168,1.52131576
-5.6,14954,114087,84662,26777,1.48210864,1.81519552,1.71632752,1.52183392
-5.7,15036,115199,86780,27022,1.49248604,1.82903372,1.73354588,1.532759
-5.8,15283,115212,86109,27365,1.49312284,1.82888428,1.7310982,1.53371836
-6,15367,114946,86017,27123,1.49193964,1.82652508,1.72932364,1.5314398
-6.1,15367,115246,86810,27390,1.49580204,1.83139548,1.73585052,1.53619932
-6.2,15112,114538,85637,26487,1.48631448,1.82038584,1.72327848,1.52453448
-6.3,15041,114393,85758,26339,1.4853858,1.81920852,1.72299492,1.52334708
-6.4,15309,114675,86038,27145,1.49093252,1.82480228,1.72858196,1.53070148
-6.5,15225,114126,85986,25878,1.4851066,1.81741396,1.72286356,1.52090068
-6.7,14299,105732,80013,19590,1.4207052,1.72792008,1.64150424,1.43848296
-6.8,14111,103506,78222,17708,1.40278644,1.70315364,1.6181994,1.41487236
-6.9,11862,90506,70102,14772,1.3205236,1.58476744,1.51621,1.3303012
-7,10607,82401,65649,12135,1.2695888,1.51081664,1.45452992,1.27472288
-7.1,10091,81606,65046,11718,1.261235,1.5015254,1.4458838,1.26670172
-7.2,10564,79817,65128,11600,1.2589846,1.49167468,1.44231964,1.26246556
-7.3,9822,79452,64526,10588,1.24876384,1.48272064,1.43256928,1.2513376
-7.4,9618,79132,64469,10328,1.24568996,1.479257,1.42998932,1.24807556
-7.5,9923,78798,64163,11605,1.24939004,1.48081004,1.43163644,1.25504156
-7.6,9709,78915,64297,11203,1.2476344,1.48016656,1.43105008,1.25265424
-7.7,9863,79317,64616,10909,1.24980188,1.48316732,1.43377196,1.25331644
-7.8,9834,78206,63778,10452,1.24278904,1.47251896,1.42404088,1.24486552
-7.9,9026,70723,57456,8571,1.1932312,1.40053312,1.355956,1.1917024
-8,9004,68029,55945,8115,1.17985756,1.37818156,1.33757932,1.17687052
-8.1,6771,53804,45101,4238,1.08380632,1.2418372,1.21259512,1.07529544
-8.2,6704,47415,40139,3392,1.04875144,1.1855404,1.16109304,1.03762312
-8.3,5896,44979,37195,2963,1.02724428,1.15856316,1.13240892,1.0173894
