--- intensity_force_single.csv
@@ -1,52 +1,0 @@
-intensity,force
-79852,1.35436
-62680,1.18874
-35802,0.94472
-5726,0.4361
-28720,0.95844
-44120,1.14072
-63270,1.28772
-78593,1.4112
-101200,1.52194
-123480,1.65228
-148082,1.79536
-235850,2.3177
-223087,2.13248
-233210,2.20304
-242854,2.24714
-251932,2.31084
-282010,2.57936
-298438,2.70872
-308021,2.75968
-329269,2.86748
-335778,2.88904
-347852,2.98116
-372500,3.2438
-395754,3.41628
-425903,3.69362
-7263,0.7693
-3917,0.6321
-9562,0.7693
-13125,0.7301
-22930,0.94276
-18790,0.8183
-25872,0.99078
-34318,1.13484
-52642,1.2789
-69620,1.3377
-117580,1.6023
-91432,1.33182
-108220,1.48862
-136950,1.70324
-156487,1.76596
-190984,1.96294
-213845,2.10014
-253450,2.3569
-283248,2.53722
-317250,2.77732
-325860,2.79202
-327350,2.8126
-342587,2.94196
-353980,3.04486
-385470,3.38296
-425680,3.72694
