--- intensity_force_two.csv
@@ -1,316 +1,0 @@
-intensity1,intensity2,force
-6452,3942,0.75068
-11925,9453,0.85652
-21258,18393,1.01626
-24589,22380,1.08388
-35942,32975,1.24264
-52420,49210,1.38866
-54925,55620,1.36416
-82950,83721,1.59152
-111825,109340,1.77674
-135820,132056,1.92864
-172530,160953,2.19324
-203952,191968,2.44314
-227863,217265,2.62248
-271562,246400,2.86748
-297862,264112,3.00174
-9250,14232,0.6223
-17730,25248,0.7987
-23193,32458,0.92708
-33890,43072,1.08976
-56892,63250,1.323
-72514,71930,1.43472
-118932,96956,1.77184
-179260,133860,2.205
-214956,155687,2.43334
-238757,176381,2.58622
-335412,232799,3.70244
-350907,232685,3.66814
-361389,232851,3.67794
-2430,2630,0.84378
-6135,6023,0.93688
-35230,20156,1.26028
-72103,35625,1.58172
-87356,47256,1.6562
-10049,61236,1.8424
-114320,74854,1.9943
-125513,85635,2.1315
-135484,88254,2.23048
-139159,90956,2.25596
-157400,113520,2.45098
-168926,137823,2.55976
-189653,156233,2.70872
-212456,171456,2.88904
-222156,178564,2.95176
-228465,187166,3.06054
-239564,197338,3.22322
-267406,215465,3.31436
-3252,563,0.34398
-5862,4636,0.49588
-8516,9856,0.60662
-11256,19514,0.70168
-17165,28732,0.83692
-26156,39815,0.99176
-35261,50265,1.10838
-44865,60945,1.2103
-52943,71956,1.27694
-59476,78746,1.3377
-67645,91456,1.41414
-86578,108745,1.5631
-97915,128302,1.76596
-119845,140562,1.89826
-127825,153256,2.0727
-159569,161375,2.18442
-184325,170659,2.30692
-205882,196458,2.56564
-231945,224689,2.92628
-1485,685,0.56056
-2856,1542,0.67718
-3256,1956,0.68502
-3675,3356,0.72128
-4120,3958,0.74676
-31565,25945,1.08682
-33945,27326,1.10054
-36945,30845,1.1466
-61825,55756,1.46902
-74536,75856,1.5778
-94585,101252,1.74636
-109745,128354,1.8522
-145954,177856,2.13934
-192548,222594,2.40688
-239456,252895,2.64796
-259322,267452,2.73126
-456,2189,0.49686
-2387,8245,0.49196
-2561,8840,0.52038
-4851,10453,0.5831
-6025,11785,0.5978
-9365,16589,0.6909
-11625,17925,0.7105
-11568,18526,0.71344
-13562,20598,0.75264
-31758,39456,0.8722
-32564,40956,0.90258
-40562,48256,1.01332
-48456,55465,1.0976
-53659,59452,1.127
-56985,61954,1.16424
-62945,68512,1.23186
-68545,71526,1.27596
-78456,87456,1.46216
-105423,103245,1.58172
-114320,109256,1.6317
-123456,115865,1.69932
-143265,130852,1.86396
-153265,135265,1.89336
-175265,150156,2.07956
-184521,154956,2.0972
-192564,159456,2.14816
-52956,1452,1.06036
-69945,6752,1.225
-83856,9225,1.2985
-98856,13456,1.37494
-121564,20562,1.5239
-142562,28156,1.68854
-162256,34945,1.83848
-167945,37356,1.86298
-179256,45645,2.00508
-192656,57945,2.15012
-202845,69456,2.26576
-103825,3456,1.37396
-122563,6256,1.51018
-143256,10195,1.63954
-165235,13825,1.73852
-197125,21756,1.89238
-229154,28465,2.00508
-264956,44026,2.17658
-278561,57356,2.2932
-133452,1856,1.53566
-160152,8564,1.67678
-204756,22512,1.92668
-216512,26515,1.97372
-222654,27846,1.97078
-258056,40756,2.1952
-267456,44852,2.2344
-225145,29845,1.86984
-243561,34215,1.99332
-32745,2756,1.11818
-42856,5248,1.24754
-60586,9925,1.38376
-80564,17789,1.52978
-89563,20725,1.58074
-111568,35065,1.75714
-12565,48652,1.8669
-144356,65485,2.09132
-152654,75346,2.16678
-159156,84756,2.23048
-19452,6945,1.28478
-22156,9545,1.28576
-26245,12845,1.32496
-29956,15465,1.35142
-35645,19465,1.40826
-38156,22458,1.4455
-47562,29465,1.57584
-50265,32125,1.5827
-81451,54526,1.91688
-109526,87556,2.1609
-128965,122756,2.3618
-137256,134856,2.39218
-3942,6452,0.69972
-9453,11925,0.8085
-18393,21258,0.96628
-22380,24589,1.02018
-32975,35942,1.18972
-49210,52420,1.31712
-55620,54925,1.29752
-83721,82950,1.52096
-109340,111825,1.68854
-132056,135820,1.85318
-160953,172530,2.07466
-191968,203952,2.32848
-217265,227863,2.48822
-246400,271562,2.69304
-264112,297862,2.84102
-14232,9250,0.56252
-25248,17730,0.73598
-32458,23193,0.8575
-43072,33890,1.01038
-63250,56892,1.2299
-71930,72514,1.33672
-96956,118932,1.65718
-133860,179260,2.07074
-155687,214956,2.29222
-176381,238757,2.41276
-232799,335412,3.43588
-232685,350907,3.39864
-232851,361389,3.40746
-2630,2430,0.74676
-6023,6135,0.833
-20156,35230,1.1466
-35625,72103,1.46314
-47256,87356,1.53762
-61236,10049,1.72186
-74854,114320,1.8914
-85635,125513,2.01194
-88254,135484,2.108078
-90956,139159,2.12954
-113520,157400,2.32456
-137823,168926,2.4255
-156233,189653,2.57152
-171456,212456,2.74694
-178564,222156,2.80574
-187166,228465,2.90472
-197338,239564,3.04192
-215465,267406,3.13698
-563,3252,0.44394
-4636,5862,0.5929
-9856,8516,0.686
-19514,11256,0.7889
-28732,17165,0.91924
-39815,26156,1.07212
-50265,35261,1.18482
-60945,44865,1.2691
-71956,52943,1.33574
-78746,59476,1.38572
-91456,67645,1.46804
-108745,86578,1.59642
-128302,97915,1.78654
-140562,119845,1.91296
-153256,127825,2.09818
-161375,159569,2.20696
-170659,184325,2.32064
-196458,205882,2.57348
-224689,231945,2.9155
-685,1485,0.61642
-1542,2856,0.7252
-1956,3256,0.77028
-3356,3675,0.79968
-3958,4120,0.82418
-25945,31565,1.14072
-27326,33945,1.15836
-30845,36945,1.19854
-55756,61825,1.50234
-75856,74536,1.61014
-101252,94585,1.76694
-128354,109745,1.85808
-177856,145954,2.14326
-222594,192548,2.4059
-252895,239456,2.63914
-267452,259322,2.7293
-2189,456,0.53312
-8245,2387,0.50862
-8840,2561,0.53704
-10453,4851,0.59878
-11785,6025,0.61348
-16589,9365,0.70266
-17925,11625,0.72128
-18526,11568,0.72128
-20598,13562,0.7595
-39456,31758,0.87808
-40956,32564,0.9016
-48256,40562,1.00646
-55465,48456,1.0927
-59452,53659,1.12602
-61954,56985,1.15052
-68512,62945,1.21814
-71526,68545,1.25734
-87456,78456,1.4406
-91246,91432,1.44354
-103245,105423,1.56016
-109256,114320,1.61308
-115865,123456,1.68168
-130852,143265,1.84436
-135265,153265,1.87572
-150156,175265,2.0531
-154956,184521,2.0776
-159456,192564,2.13052
-1452,52956,0.53018
-6752,69945,0.6909
-9225,83856,0.76832
-13456,98856,0.84476
-20562,121564,0.99274
-28156,142562,1.14464
-34945,162256,1.28086
-37356,167945,1.29654
-45645,179256,1.42296
-57945,192656,1.54154
-69456,202845,1.63856
-3456,103825,0.82124
-6256,122563,0.95746
-10195,143256,1.07996
-13825,165235,1.176
-21756,197125,1.32496
-28465,229154,1.43374
-44026,264956,1.58564
-57356,278561,1.69638
-1856,133452,1.03586
-8564,160152,1.18972
-22512,204756,1.4063
-26515,216512,1.4504
-27846,222654,1.44452
-40756,258056,1.65816
-44852,267456,1.6954
-29845,225145,1.35436
-34215,243561,1.47098
-2756,32745,0.75656
-5248,42856,0.90552
-9925,60586,1.05252
-17789,80564,1.20638
-20725,89563,1.2593
-35065,111568,1.4357
-48652,12565,1.54154
-65485,144356,1.74636
-75346,152654,1.8081
-84756,159156,1.86494
-6945,19452,0.87808
-9545,22156,0.8967
-12845,26245,0.9261
-15465,29956,0.96236
-19465,35645,1.0192
-22458,38156,1.05644
-29465,47562,1.18776
-32125,50265,1.1858
-54526,81451,1.4847
-87556,109526,1.69932
-122756,128965,1.88258
-134856,137256,1.911
