+++ legacy_files/camera_info_get.py
