--- multi-point_contact.avi
@@ -1,1 +1,0 @@
-
